Implementing initial plumbing for native filepicker and tidying up gui codebase
