+++ gui/constants/colors.py
@@ -1,0 +1,4 @@
+MANGO_DARK_GREEN = "#609949"
+MANGO_ORANGE = "#f3921e"
+MANGO_ORANGE_LIGHT = "#f9bc30"
+ACCENT = "white"
