+++ gui/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "gui",
+  "version": "1.0.0",
+  "description": "",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "MIT",
+  "packageManager": "pnpm@10.7.0",
+  "devDependencies": {
+    "@vitejs/plugin-vue": "^6.0.3",
+    "@vue/tsconfig": "^0.8.1",
+    "typescript": "^5.9.3",
+    "vite": "^7.3.1",
+    "vite-plugin-checker": "^0.12.0",
+    "vite-plugin-css-injected-by-js": "^3.5.2",
+    "vite-plugin-lib-inject-css": "^2.2.2",
+    "vue-tsc": "^3.2.4"
+  }
+}
