+++ gui/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "@vue/tsconfig/tsconfig.dom.json",
+  "compilerOptions": {
+    "types": ["vite/client"]
+  }
+}
